continue cleanup; use gatsby image for multiple files
