+++ src/assets/scss/detail-layout.scss
@@ -266,7 +266,7 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  align-items: start;
+  align-items: flex-start;
   background: _palette(accent4);
   min-height: 75vh;
   padding: 3em;
