--- src/components/elements/BackgroundOverlay.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const BackgroundOverlay = ({color}) => {
-  return (
-    <div className={`bg-${color} w-full h-full absolute inset-0 bg-opacity-75`} />
-  )
-}
-
-export default BackgroundOverlay
