--- src/components/elements/grid/FiveGrid.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const FiveGrid = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default FiveGrid
