+++ src/components/index.js
@@ -4,6 +4,7 @@ export { default as Contact } from './Contact'
 export { default as Footer } from './Footer'
 export { default as Hero } from './Hero'
 export { default as Layout } from './Layout'
+export { default as MediaBox } from './MediaBox'
 export { default as Menu } from './Menu'
 export { default as Nav } from './Nav'
 export { default as PageContainer } from './PageContainer'
